+++ src/main/java/services/UserServices.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import datasource.model.User;
+import exception.UserNotFoundException;
 import repository.UserRepository;
 
 public class UserServices {
@@ -21,22 +22,41 @@ public class UserServices {
 		return userList;
 	}
 
-	public User findById(Long id) {
+	public User findById(Long id) throws UserNotFoundException {
 		Optional<User> optionalUser = getOptional(id);
 		User user = null;
-
-		user = optionalUser.get();
-
+		if (!optionalUser.isPresent()) {
+			throw new UserNotFoundException("User not found by ID:" + id);
+		} else {
+			user = optionalUser.get();
+		}
 		return user;
+	}
+
+	public User authenticate(String login, String senha) throws UserNotFoundException {
+		Optional<User> optionalUser = userRepository.authenticate(login, senha);
+		User user = null;
+		
+		if (!optionalUser.isPresent()) {
+			throw new UserNotFoundException("Access Denied");
+		} else {
+			user = optionalUser.get();
+		}
+		
+		return user;		
 	}
 
 	private Optional<User> getOptional(Long id) {
 		return userRepository.findById(id);
 	}
 
-	public void delete(Long id) {
+	public void delete(Long id) throws UserNotFoundException {
 		Optional<User> optionalUser = getOptional(id);
 
-		userRepository.delete(optionalUser.get());
+		if (!optionalUser.isPresent()) {
+			throw new UserNotFoundException("User not found by ID:" + id);
+		} else {
+			userRepository.delete(optionalUser.get());
+		}
 	}
 }
